fix: add string resources , test, separate count down
