+++ faith_presentation/src/commonMain/composeResources/values/strings.xml
@@ -16,6 +16,7 @@
     <string name="copy">Copy</string>
     <string name="bismillah">Bismillah</string>
     <string name="arrow_left">Arrow left</string>
+    <string name="arrow_right">Arrow right</string>
     <string name="al_baqarah">Al-Baqarah</string>
     <string name="snack_bar_title">Success</string>
     <string name="app_name">MENA</string>
@@ -119,4 +120,9 @@
     <string name="mosque_image_description">Mosque icon</string>
     <string name="mosque_details">Mosque details</string>
     <string name="view_on_map">View on map</string>
+    <string name="icon_location">location icon</string>
+    <string name="prayer_man_icon">"prayer man icon"</string>
+    <string name="dropdown_icon">Dropdown icon</string>
+    <string name="next_prayer_time_icon">next prayer time icon</string>
+    <string name="next_prayer_in">Next prayer %1$s in:</string>
 </resources>
